requirements.txt
# Requirements without versions
pytest
ruamel.yaml

# Requirements with major version numbers
pandas >= 1

